Move Status struct to common
This commit moves the Status struct definition to the rest-api-common
crate.
